+++ app/sprinkles/core/src/ServicesProvider/ServicesProvider.php
@@ -198,6 +198,7 @@ class ServicesProvider
          */
         $container['classMapper'] = function ($c) {
             $classMapper = new ClassMapper();
+            $classMapper->setClassMapping('query_builder', 'UserFrosting\Sprinkle\Core\Database\Builder');
             $classMapper->setClassMapping('throttle', 'UserFrosting\Sprinkle\Core\Database\Models\Throttle');
             return $classMapper;
         };